requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
asgiref==3.9.1
    # via django
django==5.2.6
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
psycopg[binary]==3.2.10
    # via -r requirements.in
psycopg-binary==3.2.10
    # via psycopg
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dotenv==1.1.1
    # via -r requirements.in
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via psycopg
tzdata==2025.2
    # via
    #   django
    #   psycopg
